Make MovingObstacle position change over time
In Python, method default values are evaluated at declaration (that is,
on module import), not upon call.

As a result, MovingObstacle.getPosition was given a fixed default value
of cur_time, and all calls that did not specify cur_time returned the
same value.

This was easy to see by printing the results of multiple calls to
`/api/interop/obstacles/`, and observing the same values returned each
time.