Enable database connection persistence
By default, Django creates a new database connection for each request.
This is an expensive operation, particularly for PostgreSQL.

CONN_MAX_AGE [1] allows threads to reuse a database connection for some
period of time, thus avoiding the expensive connect operation.

Issue #23 has more details of the performance implications of this
change, but it reduces the average response time from 20ms to 8ms, and
approximately doubles the maximum load the server can handle.

1: https://docs.djangoproject.com/en/1.8/ref/settings/#std:setting-CONN_MAX_AGE