Prefetch AerialPosition and GpsPosition
Automatically prefetch AerialPosition and GpsPosition for
UasTelemetry.getAccessLogForUser.  Almost every user of
getAccessLogForUser will access these fields, so fetching them up front
avoids excessive database queries.

It is worth noting that this means if these database entries change
after the UasTelemetry is fetched, but before AerialPosition is
accessed, the changes will not be reflected.  This should not be a
problem for existing use cases.

In a test database with ~115,000 UasTelemetry entries, this reduces the
response time of `/auvsi_admin/export_data.kml` from ~3min to 5-8s.  It
reduces the response time of `/auvsi_admin/evaluate_teams.csv` from
~3min to 15-30s.

References #65