Refactor AccessLog (part 1)
Refactor all of AccessLog except for getAccessLogRates.

Rename to simply and meet PEP8 guidelines:
* getAccessLogsForUser -> by_user
* userActive -> user_active
* getAccessLogsForUserByTimePeriod -> by_time_period

by_time_period is vastly simplified by simply performing a database
lookup for each time period. No longer to callers need to do a lookup
before even calling this function.

by_time_period now returns a Django QuerySet, which behaves like a list,
but is technically not a list.  As such, they need to be coerced into
lists for assertSequenceEqual to work properly in the tests.