Install server in a virtualenv
The Puppet python module is a bit of a disaster, and the only way to
do anything with more than one version of Python, you must do operations
inside of a virtualenv.

In preparation for Python 3 support in the client libraries, install the
server inside of a virtualenv. This is arguably what we should have done
from the beginning, with the primary downside of requiring you to
activate the virtualenv when manually interacting with the server.