Display raw ID fields for models with many instances
Certain models, such as GpsPosition and AerialPosition may have
thousands or hundreds of thousands of entries once a significant amount
of telemetry has been to the server.

By default, the admin interface will enumerate all options for these
ForeignKeys. This is infeasible when there are too many instances, so
just display a raw ID field instead. There will be a search button next
to the box, where a specific instance can be searched for or created.

References #13