Make puppet exit with a non-zero code on failure
For some absurd reason, puppet exits with a zero exit code, even if the
configuration fails to apply. --detailed-exitcodes is needed to make it
exit non-zero:
https://docs.puppet.com/puppet/latest/reference/man/agent.html#OPTIONS

Now Travis builds should fail early when the configuration fails to
apply.

Also fix install_args, which should be a string, not a list. This never
manifested until now.