Add a frontend for target review.
Updates the backend review PUT to use the same field so it is directly
updatable by the Angular resource object.

Updates the menus for better organization of displays/tools. Updates the
help message on the dashboard.

Adds a new resource for target review requests. Adds a page to view
and update targets under review.