Support target upload using the 2017 object file format
The 2017 object file format is simply the API target JSON format, so we
can simply use the client types.

Currently there is no deduplication. Targets will be uploaded multiple
times if the tool is run multiple times. This will be a bit annoying if
some of the target files are invalid.

Cleanup some of the logger setup along the way.

Fixes 178