Add healthcheck to container, poll in Travis.
Purpose is to decrease sporatic CI test failures because the client
tests start before the server is ready. Secondary benefit to increase
health status visibility via a secondary prober.

Healthcheck which will probe the Docker container to validate that the
server is healthy. It simply polls the homepage (/) using curl. It
stores the status and any errors, and can be queried using `docker
inspect`. Format is defined here:
https://docs.docker.com/engine/reference/builder/#/healthcheck