Update waypoint scoring to 2017 rules.
In previous rules, you had to travel within 100ft and it was assumed
that you couldn't skip between waypoints as a result. In the new rules,
the constraint is only that they waypoint hits used must be in sequence,
and the max scoring sequence is used.

This new approach uses dynamic programming to find the best sequence. It
first filters the telemetry logs for the hits (much fewer) and then
calculates the max-scoring sequence recurrence.