Deprecate WebGL map. Deprecate APIs. Cleanup UI.
Eliminates the WebGL map to reduce maintenance burden. The upgraded KML
views are more featureful and easier to maintain.

Deprecates APIs which were only used in the WebGL map as they are no
longer used.

Cleans up the UI to contain the viewport. Resizes and rearranges
components so it's easier to view.