Use protos for ODLC API and admin review interface.
Protos don't have a tri-state field (specified with value, specified
with null, or not specified). As such, this change drops the use of the
specified with null to indicate clearing of fields. Subsequently, all
PUT of ODLCS now overrides all fields with the exact contents of the
message. This is likely a simpler and more standard API anyhow.

Drops the --team_id and --actionable_override args for the interop CLI.
These are admin only fields which shouldn't be necessary in a
non-failure situation now that interop upload is required. Removing this
functionality simplifies the API.

Deletes legacy code where spotted.