Require mission ID in request for mission details.
Updates API to get all missions to be admin only, requiring all teams to
specify the mission ID to request details. The competition is moving to
multiple simultaneous missions, so now the list of missions may have
more than one. To prevent likely bugs like always selecting the first in
the list, we force teams to provide the ID of their mission.

Updates client library and CLI to reflect this change.