Fix single-team eval field names. Fix dash perf.
Single team eval was broken by the proto change when field names were
changed. This fixes the field naming and adds back the id field.

Fixes a performance issue in the dashboard where intervals weren't
stopped on page change and new intervals would be created on navigation
back. This would cause N intervals to be running, growing load on the
server by N times. Now the controller listens for destroy and cancels
the interval.