Add mission to takeoff event. Reorganize KML.
Adds the mission a TakeoffOrLandingEvent is marking for a team. This
enables telemetry to be scored for a specific mission, further enabling
the server to support multiple missions over time.

Adds filtering of users which don't have a flight in the scoring and KML
export. This will help manage multiple missions for a single competition
by not outputing two scores for a given team, one of which is noise.

Reorganizes all KML into the views folder, removes extra testing of it.
Adds additional smoke test for Live KML update handler.