Add university name to evaluation out.
This will make it easier to join the interop output with the other
spreadsheets, making it easier to accurately form final scores and
feedback handouts.

Uses 'first_name' as the team name, and 'last_name' as the university,
simply for development convenience. The future bulk creation script will
use the same fields, making it effectively the same.

Removes all JS tests as they've permanently diverged and are no longer
adding value.

Removes the filtering of teams with no feedback to make testing easier.