Add GPS conversion utility page.
Many of the mission parameters are determined using the string format of
GPS rather than the decimal format. Previously an offline CLI was used
to perform this conversion. To save the interop operator from knowing
how to use a command line and CLI, this pushes it into the web
interface.