Make /api/teams public. Add age & ID. Add to CLI.
Makes the endpoint /api/teams public so that teams can use it for moving
obstacle avoidance.

Adds telemetry age so teams can extrapolate where the UAS is currently.

Adds telemetry ID so teams can dedupe the same telemetry message which
might otherwise indicate the UAS isn't moving.

Adds to the CLI so it's convenient for teams to access.